Rename vcm/config.py to vcm/utils.py
